backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
protobuf>=5.26.1
